requirements.txt
requests
beautifulsoup4
python-dotenv
Jinja2
